Using different ports and container names for MySQL and PostgreSQL.
Also avoiding a failure when deleting the Docker network if a container
is still running.

+++ _bin/stop_db.sh
@@ -19,7 +19,13 @@ requireEnvVar "DB_NETWORK_NAME"
 NETWORK_EXISTS=$(docker network ls --quiet --filter "name=${DB_NETWORK_NAME}")
 if [ -z "${NETWORK_EXISTS}" ]; then
   echo "Network ${DB_NETWORK_NAME} is not currently running."
-else
+  exit
+fi
+
+NETWORK_CONTAINERS=$(docker network inspect --format "{{len .Containers}}" "${DB_NETWORK_NAME}")
+if [ "${NETWORK_CONTAINERS}" = "0" ]; then
   docker network rm "${DB_NETWORK_NAME}" > /dev/null
   echo "Network ${DB_NETWORK_NAME} stopped."
+else
+  echo "Network ${DB_NETWORK_NAME} still has ${NETWORK_CONTAINERS} running."
 fi
